requirements.txt
omegaconf
opencv-python
open3d
pyyaml
scipy
matplotlib
networkx
shapely
carla